numfmt: split implementation into modules
